Allow cancelling auto-flash for kernel/BLCP
If for some reason network is not available or connection just changed (wifi > data / data > wifi), downloading dialog might get stuck and will need a force-close.
So we show a Cancel button which will work ONLY while downloading, for safety reasons we won't allow it while running the script to get flashed.

Signed-off-by: abrahamgcc <abrahamgcc99@gmail.com>